save student

+++ app/Http/Controllers/Apis/Teacher/GroupController.php
@@ -15,7 +15,7 @@ class GroupController extends Controller
     public function index()
     {
         try {
-            $groups = Group::all();
+            $groups = Group::where("channel_id", $this->channel->id)->get();
             return successResponse(GroupResource::collection($groups));
         } catch (\Exception $e) {
             return errorResponse("error", $e);
@@ -26,8 +26,13 @@ class GroupController extends Controller
         DB::beginTransaction();
 
         try {
+            $data = $request->validated();
 
-            $group = Group::create($request->validated());
+            $data["channel_id"] = $this->channel->id;
+
+            $data["teacher_id"] = $data["teacher_id"] ?? $this->teacher->id;
+
+            $group = Group::create($data);
 
             $group->times()->createMany($request->input('times'));
 
@@ -41,17 +46,28 @@ class GroupController extends Controller
     }
 
 
-    public function show(Group $group)
+    public function show($id)
     {
         try {
+            $group = Group::where("channel_id", $this->channel->id)->where("id", $id)->first();
+            if (!$group) {
+                return errorResponse('Group not found In this Channel!', null, [], 404);
+            }
+
             return successResponse(new GroupResource($group->load('times')));
         } catch (\Exception $e) {
             return errorResponse("error", $e);
         }
     }
 
-    public function update(GroupRequest $request, Group $group)
+    public function update(GroupRequest $request, $id)
     {
+
+        $group = Group::where("channel_id", $this->channel->id)->where("id", $id)->first();
+        if (!$group) {
+            return errorResponse('Group not found In this Channel!', null, [], 404);
+        }
+
         DB::beginTransaction();
 
         try {
@@ -75,8 +91,13 @@ class GroupController extends Controller
     }
 
 
-    public function destroy(Group $group)
+    public function destroy($id)
     {
+        $group = Group::where("channel_id", $this->channel->id)->where("id", $id)->first();
+        if (!$group) {
+            return errorResponse('Group not found In this Channel!', null, [], 404);
+        }
+
         DB::beginTransaction();
         try {
             $group->times()->delete();
